Disable Already Imported warning
This warning shows up when running unittest with the --coverage option. This is due to the include now being set in the pyproject.toml, which is required, but our test-runner is importing the files already (this is why we didn't have the `include` on the coverage run command). I validated that this did not change the coverage in the report.