Hackathon: Check Compatibility Matrix
This adds a checker for the Compatibility Matrix docs. It tries to determine the minimum/maximum version of Nautobot from the Pyproject.toml (as the min/max version in the __init__.py could be going away). If you provide the --fix flag, it will add or update the line for the compatibility. This only checks the current version, it can't go back in time and check others. This also assumes that the last line is the version, which is the example, but may not be true in every repo.